isolating plotting tests to their own file, and improving the CTV perf slightly
